Move Time into its own module
